[Dynamo][8/N] Wrap itertools.* as ItertoolsVariable
