+++ torch/_dynamo/variables/misc.py
@@ -3,14 +3,13 @@ import dataclasses
 import functools
 import inspect
 import itertools
-import operator
 import sys
 import types
 from typing import Dict, List
 
 import torch._C
 import torch._numpy as tnp
-from .. import config, polyfill, variables
+from .. import config, variables
 from ..bytecode_transformation import create_call_function, create_instruction
 from ..exc import unimplemented
 from ..guards import GuardBuilder, install_guard
@@ -778,138 +777,6 @@ class SkipFilesVariable(VariableTracker):
                 value, mutable_local=MutableLocal()
             )
         elif (
-            self.value is itertools.product
-            and not kwargs
-            and all(arg.has_unpack_var_sequence(tx) for arg in args)
-        ):
-            seqs = [arg.unpack_var_sequence(tx) for arg in args]
-            items = []
-            for item in itertools.product(*seqs):
-                items.append(variables.TupleVariable(list(item)))
-            return variables.ListIteratorVariable(items, mutable_local=MutableLocal())
-        elif (
-            self.value is itertools.chain
-            and not kwargs
-            and all(arg.has_unpack_var_sequence(tx) for arg in args)
-        ):
-            seqs = [arg.unpack_var_sequence(tx) for arg in args]
-            items = []
-            for item in itertools.chain(*seqs):
-                items.append(item)
-            return variables.ListIteratorVariable(items, mutable_local=MutableLocal())
-        elif self.value is itertools.accumulate:
-            from .builtin import BuiltinVariable
-
-            if any(key not in ["initial", "func"] for key in kwargs.keys()):
-                unimplemented(
-                    "Unsupported kwargs for itertools.accumulate: "
-                    f"{','.join(set(kwargs.keys()) - {'initial', 'func'})}"
-                )
-
-            acc = kwargs.get("initial")
-
-            if len(args) in [1, 2] and args[0].has_unpack_var_sequence(tx):
-                seq = args[0].unpack_var_sequence(tx)
-
-                if "func" in kwargs and len(args) == 1:
-                    func = kwargs["func"].call_function
-                elif len(args) == 2:
-                    func = args[1].call_function
-                elif len(args) == 1:
-                    # Default to operator.add
-                    func = BuiltinVariable(operator.add).call_function
-                else:
-                    unimplemented(
-                        "itertools.accumulate can only accept one of: `func` kwarg, pos 2 arg"
-                    )
-            else:
-                unimplemented("Unsupported arguments for itertools.accumulate")
-
-            items = []
-            if acc is not None:
-                items.append(acc)
-            for item in seq:
-                if acc is None:
-                    acc = item
-                else:
-                    try:
-                        acc = func(tx, [acc, item], {})
-                    except Exception:
-                        raise unimplemented(  # noqa: TRY200
-                            f"Unexpected failure in invoking function during accumulate. Failed running func {func}({item}{acc})"
-                        )
-                items.append(acc)
-
-            return variables.ListIteratorVariable(items, mutable_local=MutableLocal())
-        elif (
-            self.value is itertools.combinations
-            and not kwargs
-            and len(args) == 2
-            and args[0].has_unpack_var_sequence(tx)
-            and args[1].is_python_constant()
-        ):
-            iterable = args[0].unpack_var_sequence(tx)
-            r = args[1].as_python_constant()
-
-            items = []
-            for item in itertools.combinations(iterable, r):
-                items.append(variables.TupleVariable(list(item)))
-            return variables.ListIteratorVariable(items, mutable_local=MutableLocal())
-        elif self.value is itertools.groupby:
-            if any(kw != "key" for kw in kwargs.keys()):
-                unimplemented(
-                    "Unsupported kwargs for itertools.groupby: "
-                    f"{','.join(set(kwargs.keys()) - {'key'})}"
-                )
-
-            def retrieve_const_key(key):
-                if isinstance(key, variables.SymNodeVariable):
-                    return key.evaluate_expr()
-                elif isinstance(key, variables.ConstantVariable):
-                    return key.as_python_constant()
-                else:
-                    raise unimplemented(
-                        "Unsupported key type for itertools.groupby: " + str(type(key))
-                    )
-
-            if len(args) == 1 and args[0].has_unpack_var_sequence(tx):
-                seq = args[0].unpack_var_sequence(tx)
-                keyfunc = (
-                    (
-                        lambda x: (
-                            retrieve_const_key(
-                                kwargs.get("key").call_function(tx, [x], {})
-                            )
-                        )
-                    )
-                    if "key" in kwargs
-                    else None
-                )
-            else:
-                unimplemented("Unsupported arguments for itertools.groupby")
-
-            result = []
-            try:
-                for k, v in itertools.groupby(seq, key=keyfunc):
-                    result.append(
-                        variables.TupleVariable(
-                            [
-                                variables.ConstantVariable.create(k)
-                                if variables.ConstantVariable.is_literal(k)
-                                else k,
-                                variables.ListIteratorVariable(
-                                    list(v), mutable_local=MutableLocal()
-                                ),
-                            ],
-                            mutable_local=MutableLocal(),
-                        )
-                    )
-            except Exception:
-                raise unimplemented(  # noqa: TRY200
-                    "Unexpected failure when calling itertools.groupby"
-                )
-            return variables.ListIteratorVariable(result, mutable_local=MutableLocal())
-        elif (
             self.value is functools.wraps
             and not kwargs
             and len(args) == 1
@@ -947,21 +814,6 @@ class SkipFilesVariable(VariableTracker):
             return variables.functions.FunctoolsPartialVariable(
                 fn, args=rest_args, keywords=kwargs
             )
-        elif self.value is itertools.repeat:
-            if len(args) < 2:
-                return variables.RepeatIteratorVariable(
-                    *args, mutable_local=MutableLocal()
-                )
-
-            from .builder import SourcelessBuilder
-
-            return tx.inline_user_function_return(
-                SourcelessBuilder()(tx, polyfill.repeat), args, kwargs
-            )
-        elif self.value is itertools.count:
-            return variables.CountIteratorVariable(*args, mutable_local=MutableLocal())
-        elif self.value is itertools.cycle:
-            return variables.CycleIteratorVariable(*args, mutable_local=MutableLocal())
         else:
             try:
                 path = inspect.getfile(self.value)
